fix(frontend): move PRE_SURVEY_FIELDS out of pages/Demo into components/survey
DemoTab.tsx (components/admin) imported PRE_SURVEY_FIELDS from pages/Demo,
a component-to-page import inversion that drags the entire Demo route
module (and its import graph) into any bundle that pulls in DemoTab —
including the admin chunk. Move the array into
components/survey/preSurveyFields.ts, matching the existing sibling
postSurveyFields.ts, and re-point both consumers (DemoTab.tsx and
Demo.tsx itself). Array copied verbatim, diffed byte-identical against
the original.
