+++ frontend/src/components/admin/DemoTab.tsx
@@ -16,7 +16,7 @@ import * as supportApi from '../../api/support'
 import type { SupportTicket, SupportTicketSummary } from '../../types/support'
 import type { DemoAdminStats, DemoApplication as DemoApp, PostExperienceResponseAdmin } from '../../types/demo'
 import { POST_SURVEY_FIELDS } from '../survey/postSurveyFields'
-import { PRE_SURVEY_FIELDS } from '../../pages/Demo'
+import { PRE_SURVEY_FIELDS } from '../survey/preSurveyFields'
 import { SurveyFieldRenderer } from '../survey/SurveyFieldRenderer'
 import { downloadCSV, formatDate } from './shared/format'
 import { SearchInput } from './shared/primitives'
